+++ packages/components/src/components.d.ts
@@ -360,6 +360,10 @@ export namespace Components {
     }
     interface MiLocationInfo {
         /**
+          * @description The word used for "Level" when showing level info. Default is "Level".
+         */
+        "level": string;
+        /**
           * @description Location object.
          */
         "location": any;
@@ -719,7 +723,27 @@ export namespace Components {
          */
         "unit": UnitSystem;
     }
+    interface MiRouteInstructionsManeuverLegacy {
+        /**
+          * Maneuver to display given as stringified JSON.
+          * @type {string} - Maneuver/substep object passed as stringified JSON.
+         */
+        "maneuver": string;
+        /**
+          * Object with translation strings as stringified JSON.
+         */
+        "translations": string;
+        /**
+          * Set imperial or metric as default unit system.
+          * @type {UnitSystem} 'imperial' or 'metric'
+         */
+        "unit": UnitSystem;
+    }
     interface MiRouteInstructionsStep {
+        /**
+          * Programmatically close the substeps.
+         */
+        "closeSubsteps": () => Promise<void>;
         /**
           * The final step when the user has arrived to the destination location.
           * @type {string}
@@ -745,6 +769,10 @@ export namespace Components {
           * @type {boolean}
          */
         "hideIndoorSubsteps": boolean;
+        /**
+          * Programmatically open the substeps.
+         */
+        "openSubsteps": () => Promise<void>;
         /**
           * The step data object to render from. Must be passed as stringified JSON.
          */
@@ -1186,6 +1214,12 @@ declare global {
         prototype: HTMLMiRouteInstructionsManeuverElement;
         new (): HTMLMiRouteInstructionsManeuverElement;
     };
+    interface HTMLMiRouteInstructionsManeuverLegacyElement extends Components.MiRouteInstructionsManeuverLegacy, HTMLStencilElement {
+    }
+    var HTMLMiRouteInstructionsManeuverLegacyElement: {
+        prototype: HTMLMiRouteInstructionsManeuverLegacyElement;
+        new (): HTMLMiRouteInstructionsManeuverLegacyElement;
+    };
     interface HTMLMiRouteInstructionsStepElement extends Components.MiRouteInstructionsStep, HTMLStencilElement {
     }
     var HTMLMiRouteInstructionsStepElement: {
@@ -1278,6 +1312,7 @@ declare global {
         "mi-route-instructions": HTMLMiRouteInstructionsElement;
         "mi-route-instructions-heading": HTMLMiRouteInstructionsHeadingElement;
         "mi-route-instructions-maneuver": HTMLMiRouteInstructionsManeuverElement;
+        "mi-route-instructions-maneuver-legacy": HTMLMiRouteInstructionsManeuverLegacyElement;
         "mi-route-instructions-step": HTMLMiRouteInstructionsStepElement;
         "mi-route-instructions-step-legacy": HTMLMiRouteInstructionsStepLegacyElement;
         "mi-scroll-buttons": HTMLMiScrollButtonsElement;
@@ -1699,6 +1734,10 @@ declare namespace LocalJSX {
     }
     interface MiLocationInfo {
         /**
+          * @description The word used for "Level" when showing level info. Default is "Level".
+         */
+        "level"?: string;
+        /**
           * @description Location object.
          */
         "location"?: any;
@@ -2006,6 +2045,22 @@ declare namespace LocalJSX {
          */
         "unit"?: UnitSystem;
     }
+    interface MiRouteInstructionsManeuverLegacy {
+        /**
+          * Maneuver to display given as stringified JSON.
+          * @type {string} - Maneuver/substep object passed as stringified JSON.
+         */
+        "maneuver"?: string;
+        /**
+          * Object with translation strings as stringified JSON.
+         */
+        "translations"?: string;
+        /**
+          * Set imperial or metric as default unit system.
+          * @type {UnitSystem} 'imperial' or 'metric'
+         */
+        "unit"?: UnitSystem;
+    }
     interface MiRouteInstructionsStep {
         /**
           * The final step when the user has arrived to the destination location.
@@ -2039,6 +2094,10 @@ declare namespace LocalJSX {
           * @properties Object
          */
         "onStepClicked"?: (event: MiRouteInstructionsStepCustomEvent<object>) => void;
+        /**
+          * Event emitted when substeps are toggled.
+         */
+        "onSubstepsToggled"?: (event: MiRouteInstructionsStepCustomEvent<void>) => void;
         /**
           * The step data object to render from. Must be passed as stringified JSON.
          */
@@ -2306,6 +2365,7 @@ declare namespace LocalJSX {
         "mi-route-instructions": MiRouteInstructions;
         "mi-route-instructions-heading": MiRouteInstructionsHeading;
         "mi-route-instructions-maneuver": MiRouteInstructionsManeuver;
+        "mi-route-instructions-maneuver-legacy": MiRouteInstructionsManeuverLegacy;
         "mi-route-instructions-step": MiRouteInstructionsStep;
         "mi-route-instructions-step-legacy": MiRouteInstructionsStepLegacy;
         "mi-scroll-buttons": MiScrollButtons;
@@ -2348,6 +2408,7 @@ declare module "@stencil/core" {
             "mi-route-instructions": LocalJSX.MiRouteInstructions & JSXBase.HTMLAttributes<HTMLMiRouteInstructionsElement>;
             "mi-route-instructions-heading": LocalJSX.MiRouteInstructionsHeading & JSXBase.HTMLAttributes<HTMLMiRouteInstructionsHeadingElement>;
             "mi-route-instructions-maneuver": LocalJSX.MiRouteInstructionsManeuver & JSXBase.HTMLAttributes<HTMLMiRouteInstructionsManeuverElement>;
+            "mi-route-instructions-maneuver-legacy": LocalJSX.MiRouteInstructionsManeuverLegacy & JSXBase.HTMLAttributes<HTMLMiRouteInstructionsManeuverLegacyElement>;
             "mi-route-instructions-step": LocalJSX.MiRouteInstructionsStep & JSXBase.HTMLAttributes<HTMLMiRouteInstructionsStepElement>;
             "mi-route-instructions-step-legacy": LocalJSX.MiRouteInstructionsStepLegacy & JSXBase.HTMLAttributes<HTMLMiRouteInstructionsStepLegacyElement>;
             "mi-scroll-buttons": LocalJSX.MiScrollButtons & JSXBase.HTMLAttributes<HTMLMiScrollButtonsElement>;